added logic for Hamburgher menu | updated UI | moved usefull files & deleted useless files
