--- frontend/src/components/navbar.css
@@ -1,30 +1,0 @@
-.navbar {
-     background-color: #333;
-     color: white;
-     padding: 10px;
-     display: flex;
-     justify-content: space-between;
-     align-items: center;
-   }
-
-   .navbar-brand a {
-     color: white;
-     text-decoration: none;
-     font-weight: bold;
-   }
-
-   .navbar-links {
-     list-style: none;
-     padding: 0;
-     margin: 0;
-     display: flex;
-   }
-
-   .navbar-links li {
-     margin-left: 15px;
-   }
-
-   .navbar-links a {
-     color: white;
-     text-decoration: none;
-   }
