--- frontend/src/components/navbar.tsx
@@ -1,12 +1,0 @@
-  
-   import './navbar.css'; // Importa il foglio di stile (opzionale)
-
-   const Navbar = () => {
-     return (
-       <nav className="navbar">
-        
-       </nav>
-     );
-   };
-
-   export default Navbar;
