+++ frontend/src/components/ui/HamMenu.tsx
@@ -1,0 +1,9 @@
+
+export default function Ham(){
+
+    return (
+        <svg className="h-6 w-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
+            </svg>
+    )
+}
